+++ PCF8574.h
@@ -2,7 +2,7 @@
  * PCF8574 GPIO Port Expand
  *
  * AUTHOR:  Renzo Mischianti
- * VERSION: 2.3.2
+ * VERSION: 2.3.3
  *
  * https://www.mischianti.org/2019/01/02/pcf8574-i2c-digital-i-o-expander-fast-easy-usage/
  *
@@ -110,17 +110,17 @@ public:
 	PCF8574(uint8_t address, uint8_t interruptPin,  void (*interruptFunction)() );
 
 #if !defined(__AVR) && !defined(ARDUINO_ARCH_SAMD) && !defined(__STM32F1__) && !defined(TEENSYDUINO)
-	PCF8574(uint8_t address, uint8_t sda, uint8_t scl);
-	PCF8574(uint8_t address, uint8_t sda, uint8_t scl, uint8_t interruptPin,  void (*interruptFunction)());
+	PCF8574(uint8_t address, int sda, int scl);
+	PCF8574(uint8_t address, int sda, int scl, uint8_t interruptPin,  void (*interruptFunction)());
 #endif
 
 #ifdef ESP32
 	///// changes for second i2c bus
 	PCF8574(TwoWire *pWire, uint8_t address);
-	PCF8574(TwoWire *pWire, uint8_t address, uint8_t sda, uint8_t scl);
+	PCF8574(TwoWire *pWire, uint8_t address, int sda, int scl);
 
 	PCF8574(TwoWire *pWire, uint8_t address, uint8_t interruptPin,  void (*interruptFunction)() );
-	PCF8574(TwoWire *pWire, uint8_t address, uint8_t sda, uint8_t scl, uint8_t interruptPin,  void (*interruptFunction)());
+	PCF8574(TwoWire *pWire, uint8_t address, int sda, int scl, uint8_t interruptPin,  void (*interruptFunction)());
 #endif
 
 	bool begin();
@@ -223,8 +223,8 @@ private:
 	#  endif
 	#endif
 
-	uint8_t _sda = DEFAULT_SDA;
-	uint8_t _scl = DEFAULT_SCL;
+	int _sda = DEFAULT_SDA;
+	int _scl = DEFAULT_SCL;
 
 	TwoWire *_wire;
